Rename Laplace_diff to LMRF
