+++ cuqi/distribution/__init__.py
@@ -6,9 +6,8 @@ from ._gamma import Gamma
 from ._gaussian import Gaussian, JointGaussianSqrtPrec
 from ._gmrf import GMRF
 from ._inverse_gamma import InverseGamma
-from ._laplace_diff import Laplace_diff
-from ._laplace import Laplace
 from ._lmrf import LMRF
+from ._laplace import Laplace
 from ._lognormal import Lognormal
 from ._normal import Normal
 from ._posterior import Posterior
