+++ demos/try10_Data_Object.py
@@ -14,7 +14,7 @@ np.random.seed(0)
 import cuqi
 from cuqi.testproblem import Deconvolution1D
 from cuqi.model import LinearModel
-from cuqi.distribution import Gaussian, Laplace_diff, Cauchy_diff
+from cuqi.distribution import Gaussian, LMRF, Cauchy_diff
 from cuqi.sampler import CWMH
 from cuqi.problem import BayesianProblem
 from cuqi.samples import Samples
