+++ tests/test_testproblem.py
@@ -150,7 +150,7 @@ def test_Abel():
 #TODO. Add tests for custom PSF
 @pytest.mark.parametrize("prior",[
     (cuqi.distribution.Gaussian(np.zeros(128**2), 1, name="x")),
-    #(cuqi.distribution.Laplace_diff(np.zeros(128**2), 1, "zeros")),
+    #(cuqi.distribution.LMRF(np.zeros(128**2), 1, "zeros")),
     #(cuqi.distribution.Cauchy_diff(np.zeros(128**2), 1, "zeros")),
 ])
 def test_Deconvolution2D_Sampling_prior(prior):
